# internal/keeper/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeper API",
    "description": "API spec for Tacito Square Keeper component, managing agent configurations, communities, and support entities.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Primary API Version 1"
    }
  ],
...
  "paths": {
...
    "/communities": {
      "post": {
        "summary": "Create a new Community logical boundary",
        "operationId": "createCommunity",
        "tags": [
          "community/communities"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created Community",
            "headers": {
              "Location": {
                "description": "The URI of the newly created communities",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all Communities",
        "operationId": "listCommunities",
        "tags": [
          "community/communities"
        ],
        "responses": {
          "200": {
            "description": "List of Communities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Community"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "name is required"
          }
        }
      },
...
      "CreateCommunityRequest": {
        "type": "object",
        "required": [
          "name",
          "topology"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "sales-agents-net"
          },
          "description": {
            "type": "string"
          },
          "topology": {
            "type": "string",
            "enum": [
              "hub-spoke"
            ]
          },
          "configuration": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "Community": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "topology",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "topology": {
            "type": "string",
            "enum": [
              "hub-spoke"
            ]
          },
          "configuration": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "active",
              "suspended",
              "terminated"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}